refactor(Product): move production code for product
