+++ src/main/java/kitchenpos/asis/application/MenuService.java
@@ -5,8 +5,8 @@ import kitchenpos.asis.domain.MenuGroup;
 import kitchenpos.asis.domain.MenuGroupRepository;
 import kitchenpos.asis.domain.MenuProduct;
 import kitchenpos.asis.domain.MenuRepository;
-import kitchenpos.asis.domain.Product;
-import kitchenpos.asis.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.asis.infra.PurgomalumClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
