+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.asis.domain;
+package kitchenpos.product.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
