+++ src/main/java/kitchenpos/product/domain/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.asis.domain;
+package kitchenpos.product.domain;
 
 import java.util.List;
 import java.util.Optional;
