+++ src/main/java/kitchenpos/product/infra/JpaProductRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.asis.domain;
+package kitchenpos.product.infra;
 
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
