+++ src/test/java/kitchenpos/asis/application/MenuServiceTest.java
@@ -4,8 +4,8 @@ import kitchenpos.asis.domain.Menu;
 import kitchenpos.asis.domain.MenuGroupRepository;
 import kitchenpos.asis.domain.MenuProduct;
 import kitchenpos.asis.domain.MenuRepository;
-import kitchenpos.asis.domain.Product;
-import kitchenpos.asis.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.asis.infra.PurgomalumClient;
 import kitchenpos.product.InMemoryProductRepository;
 import org.junit.jupiter.api.BeforeEach;
